Partially implemented chunk data saving for storing time of day on unload

+++ src/main/java/com/gammamicroscopii/particle/FallingLeafParticle.java
@@ -64,7 +64,7 @@ public class FallingLeafParticle extends SpriteBillboardParticle {
 
 		this.velocityX +=  -0.0075 + 0.015 * random.nextFloat();
 		this.velocityZ +=  -0.0075 + 0.015 * random.nextFloat();
-		this.velocityY -= (double)this.gravityStrength;
+		this.velocityY -= this.gravityStrength;
 
 
 		if (this.onGround) {
